BASIC TIC-TAC-TOE GAME USING PYTHON
This game is created using the Python language. Use of libraries, functions, algorithms has been made to make it more interactive and better. The coding part is very basic. The algorithms are used in a way that the winner is AI or it is a draw between the human player and the AI.