RECOMMENDATION SYSTEM IN PY (MOVIE RECOMMENDATION)
This is a basic recommendation system created using Python. In this system, libraries like numpy and pandas are used to convert the data into an easily readable format. This requires a wide range of data but here, only 5 movies are taken as an example. This is the basic model of a recommendation system taking in a reality based example of a movie recommendation.